Fix: Hold Tap: add delay before release of tap
This change is copied from tyalie's zmk fork:
https://github.com/tyalie/zmk/commit/6266c1c21453132f5d03998e3df10aa4c6d49368

In tap hold, the tap behavior is often triggered by having the
corresponding key be released. As the same action is also used to
trigger a key press we have the situation that there's almost no delay
between key press and release when it reaches the host.

This can lead to issues if additional modifiers are used for the key.
e.g. under Linux with libinput the modifier is recognized but not
applied to the following key press.

Adding a 20ms delay to the release of the tap behavior key (so ~20ms
between press and release) fixes this issue.